# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
      "ApiErrorBody": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ApiErrorDetail"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorBody",
        "type": "object"
      },
      "ApiErrorCode": {
        "enum": [
          "bad_gateway",
          "bad_request",
          "conflict",
          "email_authorization_required",
          "email_insufficient_scope",
          "email_invalid_provider_response",
          "email_provider_request_failed",
          "email_rate_limited",
          "email_sync_cursor_expired",
          "email_temporarily_unavailable",
          "forbidden",
          "http_error",
          "internal_error",
          "llm_provider_invalid_response",
          "llm_provider_request_failed",
          "llm_provider_timeout",
          "llm_provider_unavailable",
          "not_found",
          "service_unavailable",
          "unauthorized",
          "validation_error"
        ],
        "title": "ApiErrorCode",
        "type": "string"
      },
      "ApiErrorDetail": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "message",
          "type"
        ],
        "title": "ApiErrorDetail",
        "type": "object"
      },
      "ApiErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
      "EmailAccountRef": {
        "description": "Provider-neutral stable reference to a connected mailbox account.",
        "properties": {
          "account_id": {
            "minLength": 1,
            "title": "Account Id",
            "type": "string"
          },
          "provider": {
            "$ref": "#/components/schemas/EmailProviderName"
          }
        },
        "required": [
          "provider",
          "account_id"
        ],
        "title": "EmailAccountRef",
        "type": "object"
      },
      "EmailAddress": {
        "description": "Provider-normalized email address plus optional display name.\n\nProvider adapters should lower-case parsed addresses, trim display names,\nand deduplicate repeated recipients before returning metadata DTOs.",
        "properties": {
          "address": {
            "minLength": 1,
            "title": "Address",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "required": [
          "address"
        ],
        "title": "EmailAddress",
        "type": "object"
      },
      "EmailAuthorizationStartResult": {
        "description": "Provider-neutral authorization URL and scope request.",
        "properties": {
          "authorization_url": {
            "minLength": 1,
            "title": "Authorization Url",
            "type": "string"
          },
          "provider": {
            "$ref": "#/components/schemas/EmailProviderName"
          },
          "requested_scopes": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Requested Scopes",
            "type": "array"
          },
          "state": {
            "minLength": 1,
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "authorization_url",
          "state",
          "requested_scopes"
        ],
        "title": "EmailAuthorizationStartResult",
        "type": "object"
      },
      "EmailConnection": {
        "description": "Stored account connection metadata without raw OAuth token material.",
        "properties": {
          "account": {
            "$ref": "#/components/schemas/EmailAccountRef"
          },
          "connected_at": {
            "format": "date-time",
            "title": "Connected At",
            "type": "string"
          },
          "credential_expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential Expires At"
          },
          "credential_ref": {
            "$ref": "#/components/schemas/SecretRef"
          },
          "display_email": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmailAddress"
              },
              {
                "type": "null"
              }
            ]
          },
          "granted_scopes": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Granted Scopes",
            "type": "array"
          },
          "reauth_required": {
            "default": false,
            "title": "Reauth Required",
            "type": "boolean"
          }
        },
        "required": [
          "account",
          "credential_ref",
          "granted_scopes",
          "connected_at"
        ],
        "title": "EmailConnection",
        "type": "object"
      },
...
      "EmailProviderName": {
        "enum": [
          "gmail"
        ],
        "title": "EmailProviderName",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SecretKind": {
        "enum": [
          "oauth_token",
          "oauth_client",
          "llm_api_key"
        ],
        "title": "SecretKind",
        "type": "string"
      },
      "SecretRef": {
        "description": "Stable non-secret identifier for a stored secret.",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/SecretKind"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9_.:-]*$",
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9_.:-]*$",
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "provider",
          "name"
        ],
        "title": "SecretRef",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Job Search Intelligence API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/connections/{provider}/{account_id}": {
      "delete": {
        "description": "Removes one stored email connection and deletes its credential from the configured secret store. Provider-side mailbox data is never touched.",
        "operationId": "disconnect_email_connection_auth_connections__provider___account_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EmailProviderName"
            }
          },
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailConnection"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Disconnect Email Connection",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/gmail": {
      "get": {
        "operationId": "gmail_auth_url_auth_gmail_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailAuthorizationStartResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "summary": "Gmail Auth Url",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/gmail/callback": {
      "get": {
        "operationId": "gmail_auth_callback_auth_gmail_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "format": "password",
              "minLength": 1,
              "title": "Code",
              "type": "string",
              "writeOnly": true
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "State",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailConnection"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Gmail Auth Callback",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}